Fix vehicle controller pivot turn bug for #7

+++ src/vehicle_controller.hpp
@@ -12,16 +12,16 @@ public:
 
 private:
 	SpeedController& speedCo;
-	// 到達目標残中心距離
-	long remainingDistance;
-	// 目標曲率[10^6mm^(-1)]
-	long targetCurvature;
-	// 移動途中速度[mm/s], [10^3rad/s]
-	int movingCenterSpeed, movingAngularSpeed;
-	// 最終目標速度[mm/s], [10^3rad/s]
-	int terminalCenterSpeed, terminalAngularSpeed;
-	// 加速中目標速度[mm/s], 角速度[10^3rad/s]
-	int centerSpeed, angularSpeed;
+	// 到達目標残中心距離(mm * ControlFreq)
+	float remainingDistance;
+	// 目標曲率[mm^(-1)]
+	float targetCurvature;
+	// 移動途中速度[mm/s], [rad/s]
+	float movingCenterSpeed, movingAngularSpeed;
+	// 最終目標速度[mm/s], [rad/s]
+	float terminalCenterSpeed, terminalAngularSpeed;
+	// 加速中目標速度[mm/s], 角速度[rad/s]
+	float centerSpeed, angularSpeed;
 };
 
 #endif /* VEHICLE_CONTROLLER_H */
